Create initial download of meteogen data.py
Initial download from meteogen. After this you will only add to the existing dataframe (csv or excel file).
The data are not in a suitable format for our work so we drop the unwanted columns as well as make each column the data from each park. Lastly we split the energy and power to make the process latter on easier (we will be working only with the energy columns).